requirements.txt
faiss-cpu>=1.8.0,<2.0
sentence-transformers>=3.0,<3.2
numpy>=1.26,<3.0
tqdm>=4.66,<5.0
rich>=13.7,<14.0
requests>=2.31,<3.0
